Correcting initial_loop_bound type

+++ src/solvers/refinement/string_refinement.h
@@ -61,7 +61,7 @@ private:
   // Base class
   typedef bv_refinementt supert;
 
-  unsigned long initial_loop_bound;
+  unsigned initial_loop_bound;
 
   string_constraint_generatort generator;
